refactor: modify code structure #42
