+++ web/config.py
@@ -1,0 +1,23 @@
+ENV: str = "development"
+USER_EMAIL: str = "test@example.com"
+DEFAULT_GEMINI_URL: str = "https://generativelanguage.googleapis.com/v1beta/models/gemini-2.0-flash:generateContent"
+
+SYSTEM_KEYWORD_RULES: dict[str, list[str]] = {
+    "엔진": ["엔진", "부조", "노킹", "진동", "소음", "터보", "부스트"],
+    "점화": ["점화", "코일", "플러그", "미스파이어", "실화"],
+    "연료": ["연료", "인젝터", "펌프", "연비", "시동지연"],
+    "냉각": ["냉각", "수온", "과열", "냉각수", "라디에이터"],
+    "배기": ["배기", "매니폴드", "촉매", "머플러", "매연"],
+    "제동": ["브레이크", "제동", "abs", "패드", "디스크"],
+    "공조": ["에어컨", "히터", "송풍", "냉방", "공조"],
+    "변속기": ["변속", "미션", "기어", "변속충격", "슬립"],
+    "전기충전": ["배터리", "충전", "알터네이터", "발전기", "전압", "크랭킹"],
+    "조향현가": ["핸들", "조향", "현가", "하체", "쇼크", "얼라인먼트", "쏠림"],
+    "시동 시스템": ["시동모터", "스타터"],
+    "바디전장": ["도어", "창문", "와이퍼", "등화", "계기판", "스마트키"],
+}
+
+CONSUMABLE_PART_KEYWORDS: list[str] = [
+    "점화플러그", "엔진오일", "오일필터", "에어필터", "캐빈필터", "에어컨필터",
+    "브레이크패드", "브레이크라이닝", "와이퍼", "냉각수", "미션오일", "부동액",
+]
