--- web/config.yml
@@ -1,2 +1,0 @@
-# Environment Setting: development / production
-environment: "development"
